+++ src/text/index.js
@@ -74,3 +74,5 @@ Object.keys(subset).forEach(k => {
     return new subset[k](m.list, this.world, this.parent);
   };
 });
+//aliases
+Text.prototype.words = Text.prototype.terms;