+++ src/world/addTags.js
@@ -1,7 +1,11 @@
-
-//
+//extend our known tagset with these new ones
 const addTags = function(tags) {
-
+  Object.keys(tags).forEach((tag) => {
+    let obj = tags[tag];
+    obj.notA = obj.notA || [];
+    obj.downward = obj.downward || [];
+    this.tags[tag] = obj;
+  });
   return tags;
 };
 module.exports = addTags;